+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/project/renderer/extension/commonmark/CommonMarkChangelogProjectRendererTestExtension.java
@@ -1,4 +1,4 @@
-package net.splitcells.website.server.project.renderer;
+package net.splitcells.website.server.project.renderer.extension.commonmark;
 
 import net.splitcells.website.server.Config;
 import org.junit.jupiter.api.Tag;
@@ -11,21 +11,21 @@ import static net.splitcells.dem.testing.TestTypes.INTEGRATION_TEST;
 import static net.splitcells.website.server.project.ProjectRenderer.projectRenderer;
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class XmlProjectRendererTestExtension {
+public class CommonMarkChangelogProjectRendererTestExtension {
     /**
      * TODO Test extension explicitly.
      */
     @Tag(INTEGRATION_TEST)
     @Test
-    public void testXmlLayout() {
+    public void testLayout() {
         final var testSubject = projectRenderer
                 ("public"
-                        , Path.of("../net.splitcells.dem")
+                        , Path.of("../..")
                         , Path.of("../net.splitcells.website.default.content/src/main/xsl/net/splitcells/website/den/translation/to/html/")
                         , Path.of("net.splitcells.website.default.content/src/main/resources/html")
-                        , "/net/splitcells/dem"
+                        , "/net/splitcells/network"
                         , a -> Optional.empty()
                         , Config.create());
-        assertThat(testSubject.projectPaths()).contains(Path.of("net/splitcells/dem/history.html"));
+        assertThat(testSubject.projectPaths()).contains(Path.of("net/splitcells/network/CHANGELOG.html"));
     }
 }